+++ app/routes/components/IconButton/changelog.codegen.ts
@@ -1,4 +1,5 @@
 export const changelog = ''
   + '- 2025-09-09: Initial version.\n'
   + '- 2025-09-19: Support the prop `size` (`sm`, `md` and `lg`).\n'
+  + '- 2025-09-26: Use theme tokens for styling.\n'
   ;