+++ src/components/IconButton/CHANGELOG.md
@@ -1,2 +1,3 @@
 - 2025-09-09: Initial version.
 - 2025-09-19: Support the prop `size` (`sm`, `md` and `lg`).
+- 2025-09-26: Use theme tokens for styling.